Fix determination of java version
The script requires the java version to be on
line one of the output of `java -version`.
If you use _JAVA_OPTIONS to inject parameters
to each JVM, that will produce an output before
each VM is started, and the script won't be able
to find the java version, since `head -n 1` now
returns the wrong line. Grep for 'version' instead
to always get the right line from the output.

Change-Id: Iaf2798ba805d1b289b7e82239113b99479250dd8
Signed-off-by: Tim Roes <mail@timroes.de>

Conflicts:
	core/main.mk